requirements.txt
openai >= 1.3.2
python-dotenv >= 1.0.0